Allow this client to be Node package
This commit initialises the JavaScript client so that it can be
installed as a package using "npm install". This is useful for other
projects that use NodeJS or NPM and depend on this client. Specific
functionality can be exported in the "index.js" file, which currently
allows access to the UFDLServerContext class.