Complete functionality to be used by the UFDL web frontend
This commit adds/fixes missing functionality to be in line with the
current backend version and the new UFDL web frontend. This is still
currently a work in progress and still needs signifcant testing. This
version, however, is at a point where it is ready to be used in a basic
capacity with the web frontend.