Complete functionality (#2)
* Allow this client to be Node package

This commit initialises the JavaScript client so that it can be
installed as a package using "npm install". This is useful for other
projects that use NodeJS or NPM and depend on this client. Specific
functionality can be exported in the "index.js" file, which currently
allows access to the UFDLServerContext class.

* Complete functionality to be used by the UFDL web frontend

This commit adds/fixes missing functionality to be in line with the
current backend version and the new UFDL web frontend. This is still
currently a work in progress and still needs significant testing. This
version, however, is at a point where it is ready to be used in a basic
capacity with the web frontend.

* Added missing files

This commit adds a few missing files from the previous commit.

* Added comments to the index.js file

This commit added comments to the index.js file to better indicate the
functionality used by the web frontend.